Move RBAC related classes underneath RBAC project.
